worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Video processing
yt-dlp==2024.12.13
ffmpeg-python==0.2.0

# AI services
groq==0.11.0
google-generativeai==0.8.3

# Storage
boto3==1.35.72
botocore==1.35.72

# Database
supabase==2.9.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.17
httpx==0.27.2
